Move reentrancy config
