+++ dapr/actor/runtime/config.py
@@ -8,7 +8,30 @@ Licensed under the MIT License.
 from datetime import timedelta
 from typing import Any, Dict, List, Optional
 
-from dapr.actor.runtime.reentrancy_config import ActorReentrancyConfig
+
+class ActorReentrancyConfig:
+    def __init__(
+            self,
+            enabled: Optional[bool] = False,
+            maxStackDepth: Optional[int] = 32):
+        """Inits :class:`ActorReentrancyConfig` to optionally configure actor
+        reentrancy.
+
+        Args:
+            enabled (bool): Set to enable or disable reentrancy.
+            maxStackDepth (int): Limit for the number of concurrent reentrant requests
+                to an actor, further requests are denied.
+        """
+
+        self._enabled = enabled
+        self._maxStackDepth = maxStackDepth
+
+    def as_dict(self) -> Dict[str, Any]:
+        """Returns ActorReentrancyConfig as a dict."""
+        return {
+            'enabled': self._enabled,
+            'maxStackDepth': self._maxStackDepth,
+        }
 
 
 class ActorRuntimeConfig:
